+++ .github/actions/build-packages/action.yml
@@ -31,7 +31,7 @@ runs:
 
     - name: Build Packages
       run: |
-        make -C packages/integration-tests build-debug
+        make -C tests/integration build-debug
         make -C packages/template-manager build-debug
         make -C packages/orchestrator build-debug
         make -C packages/api build-debug
