+++ .github/actions/start-services/action.yml
@@ -24,7 +24,7 @@ runs:
         echo "TESTS_E2B_API_KEY=${TESTS_E2B_API_KEY}" >> .env.test
         set -x
         make migrate
-        make -C packages/integration-tests seed
+        make -C tests/integration seed
       shell: bash
 
     - name: Start Services
