+++ tests/integration/internal/tests/api/sandbox_test.go
@@ -7,8 +7,8 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/e2b-dev/infra/packages/integration-tests/internal/api"
-	"github.com/e2b-dev/infra/packages/integration-tests/internal/setup"
+	"github.com/e2b-dev/infra/tests/integration/internal/api"
+	"github.com/e2b-dev/infra/tests/integration/internal/setup"
 
 	"github.com/stretchr/testify/assert"
 )
